refactor: rename IotaSDKFFI to CIotaSDK
