+++ bindings/swift/Sources/CIotaSDK/module.modulemap
@@ -1,5 +1,5 @@
-module IotaSDKFFI {
-    header "IotaSDKFFI.h"
+module CIotaSDK {
+    header "CIotaSDK.h"
     export *
     use "Darwin"
     use "_Builtin_stdbool"
