+++ bindings/swift/uniffi.toml
@@ -1,2 +1,3 @@
 [bindings.swift]
 module_name = "IotaSDK"
+ffi_module_name = "CIotaSDK"
